station_manager.{h|cpp}: add locking.
A SIGSEGV was seen as a result of two threads accessing a station object
via the station_manager simultaneously.

Thread one modified the internals of a station object, while thread two
derefenced an iterator to the same station object to return a copy of
the object by value.

Since we returned a station by value in thread two, while the copy was
being created on the stack, a vector internal to the station was still
being modified by thread one. Bad!

Add RAII locking to all station_manager APIs that operate on stations.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>