README.md: update usage so that it matches actual arguments
Update the arguments in README.md so that they match the ones from the
usage line.
In the usage line, change "device" to "interface_name" (which was used
in the README so far), to make it clearer it has to be an actual
interface name.

While at it, add the missing newline at the end of the README.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>