main: exit on pcap loop failure
When the pcap loop stops for any reason (e.g. the interface is brought
down, or deleted), there is no recovery for it, so only cached results
will be available until station-sniffer is restarted.

This is problematic, especially because there is no way for an
external process to know that it happened.

Change the behavior so that when the pcap loop stops, station-sniffer
exits entirely. This allows a process manager (e.g. procd on OpenWrt)
to restart it (after re-creating the missing interface for example).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>