main: get the bandwidth from the wiphy
In some cases, it seems like the interface we watch does not give us
bandwidth information.

To get the bandwidth anyway, we currently try to finding another
interface with the same MAC address. This is problematic, as we don't
have much control (by default) over the MAC addresses that
OpenWrt/prplOS assigns, which means we might end up matching a station
interface for example (which still might not have bandwidth
information).

An alternative to matching the MAC address is iterate all the
interfaces and take the bandwidth from the first AP interface that
sits on the same wiphy.

For that, the wiphy is saved when we first get the information about
the interface. It is later used to match the right wiphy when
iterating them.

Note that this also fixes a bug: up to now the bandwidth was only
copied to 'measurement_radio_info' if the interface initially did not
have any bandwidth information. We now do the copy in all cases.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>