main: check for bandwidth information periodically
It's entirely possible that when station-sniffer starts up, no beacon
information is available, but that it becomes available later on.  An
example of such a situation is when a prplMesh agent has a wireless
backhaul interface defined, but is not onboarded yet. In such a case,
the APs of the device will only start beaconing once the agent is
onboarded to the controller. Until then no beacon information will be
available.

To avoid the need to restart station-sniffer in such a case, start a
new thread to check for bandwidth information periodically if it was
not available initially. Once the agent onboards, the bandwidth
information will become available.

Note that we also slightly change what was done in commit
0625732af38dbf5e8f91cf79d7676c51d0417d28 so that we stop running if
the pcap thread stops, but not if the bandwidth thread stops (in case
it happens to be the one we're joining).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>