station_manager: fix broadcast capture bug
If a broadcast MAC address was registered as a staiton of interest to
the station manager, and a non-broadcast MAC address was subsequently
registered, the station manager would stop capturing broadcast traffic
(i.e. every packet that it can hear in the radio's current band).

This is a bug that was revealed by adding unit tests -- the manager
instance should continue to capture broadcast traffic until the
broadcast MAC is explicitly removed from it's MAC table.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>