Add station association state tracking
Exposes new message type that allows a Unix client to ask if a given STA
MAC has disassociated from any BSS on the band that station-sniffer is
sniffing on.

We simply parse every packet and check if it's a management frame, and
if so:
    1. check it's subtype
    2. if disassociation, enqueue a message for clients

This allows other client applications a new way to ask about station
association state where it otherwise may not be valid (for example, a
station is not known to hostapd on a destination access point after a
virtual BSS roam occurs).

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>