socket_server.cpp: add new UDS clients outside of iteration
Calling push_back() [or any member that may cause reallocation on an STL
data structure] while iterating the data structure via a range-based
for-loop may lead to undefined behavior.

Since range-based for-loops depend on the .begin() and .end() iterators
of iterable data structures, a reallocation will cause the end()
iterator to move, leaving the previous reference used for terminating
the for-loop incorrect.

To avoid this, mark new clients that have 'rung' our server socket for
addition after iterating the current poll FD set.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>