messages.{h|cpp}: add new message types
MSG_GET_TRAFFIC_STATS gets outbound/inbound traffic statistics in
time-bounded buckets for an individual station.

MSG_GET_TRAFFIC_STATS_BUCKET_SIZE gets the width of the bucket for an
individual station

MSG_ADJUST_TRAFFIC_STATS_BUCKET_SIZE allows for the width of the capture
bucket for an individual station to be modified.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>